requirements.txt
Django==1.11.5
django-webpack-loader==0.6.0
djangorestframework==3.8.2
mysqlclient==1.3.13
Pillow==5.2.0
pytz==2018.5
